+++ mex/backend/graph/connector.py
@@ -12,7 +12,13 @@ from neo4j import (
 from neo4j.exceptions import ConstraintError, Neo4jError
 
 from mex.backend.graph.exceptions import DeletionFailedError, IngestionError
-from mex.backend.graph.models import IngestData, MExPrimarySource, Result
+from mex.backend.graph.models import (
+    MEX_EDITOR_PRIMARY_SOURCE,
+    MEX_PRIMARY_SOURCE,
+    ExtractedPrimarySourceWithHardcodedIdentifiers,
+    IngestData,
+    Result,
+)
 from mex.backend.graph.query import Query, QueryBuilder
 from mex.backend.graph.transform import (
     expand_references_in_search_result,
@@ -33,6 +39,7 @@ from mex.common.logging import logger
 from mex.common.models import (
     EXTRACTED_MODEL_CLASSES_BY_NAME,
     MERGED_MODEL_CLASSES_BY_NAME,
+    MEX_EDITOR_PRIMARY_SOURCE_STABLE_TARGET_ID,
     RULE_MODEL_CLASSES_BY_NAME,
     AnyExtractedModel,
     AnyRuleModel,
@@ -133,8 +140,8 @@ class GraphConnector(BaseConnector):
 
     def _seed_data(self) -> None:
         """Ensure the primary source `mex` is seeded and linked to itself."""
-        deque(self.ingest_items([MExPrimarySource()]))
-        logger.info("seeded mex primary source")
+        deque(self.ingest_items([MEX_PRIMARY_SOURCE, MEX_EDITOR_PRIMARY_SOURCE]))
+        logger.info("seeded primary sources 'mex' and 'mex-editor'")
 
     def close(self) -> None:
         """Close the connector's underlying requests session."""
@@ -166,7 +173,7 @@ class GraphConnector(BaseConnector):
         identifier: str | None,
         stable_target_id: str | None,
         entity_type: Sequence[str],
-        referenced_identifiers: Sequence[str] | None,
+        referenced_identifiers: Sequence[str | None] | None,
         reference_field: str | None,
         skip: int,
         limit: int,
@@ -186,6 +193,18 @@ class GraphConnector(BaseConnector):
         Returns:
             Graph result instance
         """
+        if reference_field and reference_field not in ALL_REFERENCE_FIELD_NAMES:
+            msg = "Invalid field name."
+            raise ValueError(msg)
+        if (
+            reference_field == "hadPrimarySource"
+            and referenced_identifiers
+            and MEX_EDITOR_PRIMARY_SOURCE_STABLE_TARGET_ID in referenced_identifiers
+        ):
+            referenced_identifiers = [
+                None if id_ == MEX_EDITOR_PRIMARY_SOURCE_STABLE_TARGET_ID else id_
+                for id_ in referenced_identifiers
+            ]
         query_builder = QueryBuilder.get()
         query = query_builder.fetch_extracted_or_rule_items(
             filter_by_query_string=bool(query_string),
@@ -310,11 +329,24 @@ class GraphConnector(BaseConnector):
         if reference_field and reference_field not in ALL_REFERENCE_FIELD_NAMES:
             msg = "Invalid field name."
             raise ValueError(msg)
+        filter_items_with_rules = False
+        if (
+            reference_field == "hadPrimarySource"
+            and referenced_identifiers is not None
+            and MEX_EDITOR_PRIMARY_SOURCE_STABLE_TARGET_ID in referenced_identifiers
+        ):
+            filter_items_with_rules = True
+            referenced_identifiers = [
+                id_
+                for id_ in referenced_identifiers
+                if id_ != MEX_EDITOR_PRIMARY_SOURCE_STABLE_TARGET_ID
+            ] or None
         query_builder = QueryBuilder.get()
         query = query_builder.fetch_merged_items(
             filter_by_query_string=bool(query_string),
             filter_by_identifier=bool(identifier),
             filter_by_referenced_identifiers=bool(referenced_identifiers),
+            filter_items_with_rules=filter_items_with_rules,
             reference_field=reference_field,
         )
         result = self.commit(
@@ -397,13 +429,17 @@ class GraphConnector(BaseConnector):
     def _run_ingest_in_transaction(
         self,
         tx: Transaction,
-        model: AnyExtractedModel | AnyRuleSetResponse | MExPrimarySource,
+        model: AnyExtractedModel
+        | AnyRuleSetResponse
+        | ExtractedPrimarySourceWithHardcodedIdentifiers,
     ) -> None:
         """Ingest a single item in a database transaction."""
         query_builder = QueryBuilder.get()
         if isinstance(model, AnyRuleSetResponse):
             items_to_ingest: list[
-                AnyExtractedModel | MExPrimarySource | AnyRuleModel
+                AnyExtractedModel
+                | ExtractedPrimarySourceWithHardcodedIdentifiers
+                | AnyRuleModel
             ] = [
                 model.additive,
                 model.subtractive,
@@ -431,7 +467,11 @@ class GraphConnector(BaseConnector):
 
     def ingest_items(
         self,
-        models: Iterable[AnyExtractedModel | AnyRuleSetResponse | MExPrimarySource],
+        models: Iterable[
+            AnyExtractedModel
+            | AnyRuleSetResponse
+            | ExtractedPrimarySourceWithHardcodedIdentifiers
+        ],
     ) -> Generator[None]:
         """Ingest a list of extracted models or rule set responses into the graph."""
         settings = BackendSettings.get()