+++ mex/backend/graph/transform.py
@@ -3,7 +3,11 @@ from typing import TYPE_CHECKING, TypedDict, cast
 
 from pydantic_core import ErrorDetails
 
-from mex.backend.graph.models import GraphRel, IngestData, MExPrimarySource
+from mex.backend.graph.models import (
+    ExtractedPrimarySourceWithHardcodedIdentifiers,
+    GraphRel,
+    IngestData,
+)
 from mex.common.fields import (
     FINAL_FIELDS_BY_CLASS_NAME,
     LINK_FIELDS_BY_CLASS_NAME,
@@ -44,7 +48,9 @@ def expand_references_in_search_result(
 
 
 def transform_model_into_ingest_data(
-    model: AnyExtractedModel | MExPrimarySource | AnyRuleModel,
+    model: AnyExtractedModel
+    | ExtractedPrimarySourceWithHardcodedIdentifiers
+    | AnyRuleModel,
     stable_target_id: str,
 ) -> IngestData:
     """Transform the given model into an ingestion instruction.
@@ -197,7 +203,7 @@ def validate_ingested_data(
 def get_error_details_from_neo4j_error(
     input_: AnyExtractedModel
     | AnyRuleSetResponse
-    | MExPrimarySource
+    | ExtractedPrimarySourceWithHardcodedIdentifiers
     | AnyPrimitiveType,
     error: Neo4jError,
 ) -> list[ErrorDetails]: