+++ tests/ingest/test_main.py
@@ -94,6 +94,12 @@ def test_ingest_extracted(
             "end": "00000000000000",
         },
         {
+            "end": "00000000000000",
+            "label": "hadPrimarySource",
+            "position": 0,
+            "start": "00000000000003",
+        },
+        {
             "position": 0,
             "start": "bFQoRhcVH5DHUq",
             "label": "hadPrimarySource",
@@ -110,6 +116,12 @@ def test_ingest_extracted(
             "start": "00000000000001",
             "label": "stableTargetId",
             "end": "00000000000000",
+        },
+        {
+            "end": "00000000000002",
+            "label": "stableTargetId",
+            "position": 0,
+            "start": "00000000000003",
         },
         {
             "position": 0,
@@ -368,6 +380,12 @@ def test_ingest_extracted(
             "label": "ExtractedPrimarySource",
             "identifier": "00000000000001",
         },
+        {"identifier": "00000000000002", "label": "MergedPrimarySource"},
+        {
+            "identifier": "00000000000003",
+            "identifierInPrimarySource": "mex-editor",
+            "label": "ExtractedPrimarySource",
+        },
         {"label": "MergedOrganizationalUnit", "identifier": "StandaloneRule"},
         {"label": "MergedContactPoint", "identifier": "bFQoRhcVH5DHUB"},
         {"label": "MergedContactPoint", "identifier": "bFQoRhcVH5DHUD"},
@@ -495,6 +513,12 @@ def test_ingest_rule_set(
             "end": "00000000000000",
         },
         {
+            "end": "00000000000000",
+            "label": "hadPrimarySource",
+            "position": 0,
+            "start": "00000000000003",
+        },
+        {
             "position": 0,
             "start": "bFQoRhcVH5DHUq",
             "label": "hadPrimarySource",
@@ -511,6 +535,12 @@ def test_ingest_rule_set(
             "start": "00000000000001",
             "label": "stableTargetId",
             "end": "00000000000000",
+        },
+        {
+            "end": "00000000000002",
+            "label": "stableTargetId",
+            "position": 0,
+            "start": "00000000000003",
         },
         {
             "position": 0,
@@ -794,6 +824,12 @@ def test_ingest_rule_set(
             "label": "ExtractedPrimarySource",
             "identifier": "00000000000001",
         },
+        {"identifier": "00000000000002", "label": "MergedPrimarySource"},
+        {
+            "identifier": "00000000000003",
+            "identifierInPrimarySource": "mex-editor",
+            "label": "ExtractedPrimarySource",
+        },
         {"label": "MergedOrganizationalUnit", "identifier": "StandaloneRule"},
         {"label": "MergedContactPoint", "identifier": "bFQoRhcVH5DHUB"},
         {"label": "MergedContactPoint", "identifier": "bFQoRhcVH5DHUD"},
@@ -890,7 +926,7 @@ def test_ingest_artificial_data(
     assert response.status_code == status.HTTP_204_NO_CONTENT, response.text
 
     response = client_with_api_key_write_permission.get(
-        "/v0/merged-item", params={"skip": "32", "limit": "1"}
+        "/v0/merged-item", params={"skip": "33", "limit": "1"}
     )
     assert response.status_code == status.HTTP_200_OK, response.text
     assert response.json() == {
@@ -914,7 +950,7 @@ def test_ingest_artificial_data(
                 "supersededBy": None,
             }
         ],
-        "total": 42,
+        "total": 43,
     }